requirements.txt
Flask==3.0.3
numpy==2.1.0
pillow==10.4.0
pytest==8.3.2
requests==2.32.3
torch==2.4.0
transformers==4.44.2
Werkzeug==3.0.4